pyproject.toml
[project]
name = "sliit-rag-app"
version = "0.1.0"
description = "RAG-based support assistant for SLIIT powered by LangChain and FastAPI"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]",
    "uvicorn",
    "python-dotenv",
    "langchain-core",
    "langchain-community",
    "langchain-huggingface",
    "langchain-openai",
    "sentence-transformers",
    "langchain-text-splitters",
    "langchain-chroma",
    "supabase>=2.28.0",
    "pydantic[email]",
    "passlib[bcrypt]",
    "python-jose[cryptography]",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "httpx>=0.28.1",
    "email-validator>=2.3.0",
    "boto3>=1.35.0",
    "python-multipart>=0.0.6",
    "langchain-google-genai>=2.0.0",
    "pypdf>=5.0.0",
]

[tool.setuptools.packages.find]
include = ["app*"]
